[opm] - Add registry add and rm bundle commands
Adds opm registry command to add bundle content from images to the
database. For non-default bundles, once the database has been
updated, the bundle and csv field are deleted from all bundles that
are not the tip of the default channel of the package being updated.

Adds opm registry command to rm a package from the database (removing
all owned APIs).