Bundle migrations in binary
Use the go-bindata tool to generate a golang migrations file to
compile the content of the db_migrations folder into the bundle. This
ensures that the compiled binaries are portable and do not require
around the migrations themselves along with the binary.

Also updated the makefile to pull the go-bindata file and generate the
`migration.go` file as part of `make build`