Add `--mode` flag to `opm` add commands
Update `opm registry add` to take a new `--mode` flag. This allows
users to insert graph updates without explicitly defining the `replaces`
field in their CSVs. Instead, the parser takes the implicit version
field set on the CSV to determine where in the graph the new bundle
should be added and updates the channel_entry table accordingly.

The `--mode` flag takes three options: [`replaces`, `semver` and
`semver-skippatch`], and defaults to `replaces`. If replaces is
specified, the normal directed graph evaluation is used to determine
updates. If `semver` is specified, semantic version is used to determine
updates. If `semver-skippatch` is used, the patch version is ignored
and a bundle that is added with a patch version that already has an
entry will instead be inserted into the *same* place in the graph as
the existing entry.

Also wired up the `opm index add` command to pipe along the `--mode`
flag value.

This opens up the ability in the future to add an `opm registry rm`
command aimed at individual bundles