fix(load): add synthetic nodes in every channel that a replacement is in
This adds an additional test to the imageloader test suite that shows
the issue (it fails without the other changes in this commit).

The update logic was different from the "new" logic for adding bundles.
To simplify reasoning about these two paths, they're now the same -
updating will recalculate the graph from the new entrypoints and
replace whatever was there before.