[feature] CSVless bundle
This commit starts the support for csvless bundle by synthesizing a CSV based on Kube manifests in the bundle manifests folder and non-inferable information provided in the metadata folder. Amongst non-inferable information, we require a YAML file to supply most of the non-inferable information, an annotations.yaml to point to the directory names of manifests and metadata directory, and a dependencies.yaml to point out required APIs including CRDs and APIServices. Each provided APIService expects a description that contains at least the kind information in the non-inferable CSV info YAML file. This feature supports APISevrices and CRDs of all currently existing versions. This feature is provided by `generate-csv` command and will validate the bundle format and content to ensure correctness.