Make default channel optional
This commit leaves default channel blank if not specified or is not contained in channels. This affects the `annotations.yaml` and make it "" for default channels. This will not affect index.db as the code to choose a channel to be the default still lives there. It should remain there since OLM uses that value.