opm bundle extract shouldn't validate annotations
Currently, the configmap writer attempts to validate the annotations
associated with the bundle. In reality, most of these annotations are
currently unused by the time the bundle is being unpacked and applied
to the cluster. There was an issue where default channel was no longer
required, but it was still validated by extract because it has a
separate validation function.

These annotations should be validated when building and
adding to the registry (as it exists now). This commit removes
annotation validation from the configmap writer in opm alpha bundle
extract