fix(unpack): support unpacking readonly folders
layer unpacking preserves the original filemode. If the layer contains
a readonly director, this would cause unpack to fail, because the dir
on disk would not be writeable, and it's children would fail to write.

this change makes all unpacked files owner-writable, which has some
implications for how unpacked files should be used, but for purposes
within opm this will not be an issue.