fix: registry rm orphaned bundles
In some cases, bundles can be orphaned from channel graphs (for example,
a new bundle could be added as the head of the channel that does not
specify a replacement). When that happens, registry/index rm commands in
opm cannot remove these bundles because the bundles themselves do not
have an association in the database with the package directly.

This commit modifies the operatorbundle table to include the package
name of the bundle so that bundles can be directly referenced when
calling registry rm, rather than being joined on the channel_entry table
(and thus beholden to the graph itself).

This commit includes a migration to update existing bundles to associate
the package name with bundles, however in the case that the bundle has
been orphaned we have no guarantee that a bundle should be associated
with a specific package name. As such, there could be existing indexes
that still include these orphaned bundles and cannot be reassociated
automatically.