Adding deprecated property type to bundles
Adding support marking bundles as deprecated and truncating the update
graph of a given package. Deprecated versions will not be installable.
Deprecating a bundle can result in the removal of channels but this is
not permitted for the default channel. Bundles that are not in the
index will be ignored during deprecation.