Specify dependency lists
This commit introduces a complete implementation of specifying
dependencies as a set of properties for a given package. It updates and
migrates the database schema to support a more generic value and the
validator to handle that format.

Additionally, it fixes an incidental bug which caused the previous
dependency insert to duplicate rows in the dependencies table
when deps are specified in both the required api field of the CSV
and the dependency file in the metadata folder of the bundle