Bug 1868125: [opm] Add valid bundles to index when `permissive` enabled
When the `opm index add` was used with the `permissive` flag, the index
was being created, but the valid bundles from the list of bundles provided
with the command was not being added. Instead, an empty index was being
created. This PR fixes the issue by adding the valid bundles to the index,
while ignoring the invalid ones.