fix(opm-index-add) Throw error when a bundle is invalid
With the introduction of #503, when a bundle from the list of bundles
being added with the `opm index add` command was invalid, opm was not
throwing an error, and was building the index with the valid bundles
instead, even when the `permissive` mode was not enabled. This PR
fixes the issue.