Disable broken ppc64le and s390x release builds.
We don't have machines with either architecture to support native
builds since TravisCI ended its free plan for open source
projects. Since a large part of opm's functionality depends on linking
to SQLite, it needs to be built with CGO enabled, either on the target
platform or with a suitable cross-compilation toolchain.