Dockerfile: Add a symbolically linked release Dockerfile
Introduce the operator-registry.Dockerfile that's symbolically
linked from the root Dockerfile.

This is likely only needed in the short term while we iron out the
upstream/downstream split, and needed as the openshift/release
ci-operator is expecting to build the registry container image using the
operator-registry.Dockerfile (what's currently present in
openshfit/ocp-build-data for the registry 4.8 images) and that file doesn't
exist anywhere in the repository.

Signed-off-by: timflannagan <timflannagan@gmail.com>