feat(bundle): consume explicit properties
Define a new configuration file that can be provided by bundle authors to explicitly declare properties.
The file itself can have any name, but must both exist in the top-level metadata directory of the bundle
and consist of YAML with a top-level properties element containing an array of type/value tuples.

Signed-off-by: Nick Hale <njohnhale@gmail.com>