fix(opm): assume default db location in image
Currently, all opm index command operations fail on container images
that do not have the `containertools.DbLocationLabel` set. This PR
introduces a change that tries to continue the opm index operation
by assuming a default db location while printing out a warning. If
the assumption holds wrong, users will see the error when there is
an attempt to access the db from within the binary, while the warning
would serve as the accurate indicator of the missing lable.

Signed-off-by: Anik Bhattacharjee <anikbhattacharya93@gmail.com>