Set legacy API fields in ListBundles using available properties. (#724)
For each bundle in its result, ListBundles has been making three
separate database queries -- for properties, dependencies, and plural
API names -- via GetApisForEntry in order to populate the bundle
fields providedApis and requiredApis. Properties and dependencies are
already in hand and can be reused. Setting plural names on the
elements of providedApis and requiredApis is pointless because plurals
are discarded by the catalog operator at runtime.

Signed-off-by: Ben Luddy <bluddy@redhat.com>