Generate package property during migration when no provided APIs.
All entries beginning with schema version 9 should have at least one
property -- a generated package property. This did not occur for
catalog entries that had no provided APIs.

The existing migration is updated to fix the bug, and a new
migration is added to retroactively generate the package property for
databases that have already applied the existing migration.

Signed-off-by: Ben Luddy <bluddy@redhat.com>