The `--skip-tls` flag shouldn't use the http protocol
When using the `operator-sdk run bundle`'s `--skip-tls` flag which is
described as:

> skip authentication of image registry TLS certificate when pulling a bundle image in-cluster

It tries to access the image registry (given in the <bundle_image>
positional argument) using HTTP rather than HTTPS.

This behavior is unexpected and fails when the image registry only
speaks the HTTPS protocol.

This commit fixes it so that the `https` scheme is still used even when
the `--skip-tls` flag is specified

The commit (a16399f05c380ba432691d2e1d99a69aa5a269e2) which seems to have introduced
this behavior doesn't mention this behavior anywhere, so I'm assuming
it's unintentional and therefore a bug that needs fixing.

Signed-off-by: Omer Tuchfeld <otuchfel@redhat.com>