feat(validate): update validate library to validate properties
A recent request was made for us to add validation logic around
properties defined in the CSV as well as the properties.yaml file.
This commit adds this logic via warnings that trigger when a CSV
and properties.yaml share a property or either duplicates a type/value
pair among their own properties.

In order to do this, a number of changes were made:

* validate if shared properties are in both properties.yaml and CSV
* check if properties.yaml or CSV have duplicated properties
* add unit tests to confirm logic
* add testdata files for unit tests
* update TestValidateBundleFormat to be in a table format

Signed-off-by: Tyler Slaton <tyslaton@redhat.com>