Add composite veneer to `opm alpha render-veneer ...` (#1054)
* initial implementation pass

Signed-off-by: Bryce Palmer <bpalmer@redhat.com>

* initial implementation pass

Signed-off-by: Bryce Palmer <bpalmer@redhat.com>

* update implementation

Signed-off-by: Bryce Palmer <bpalmer@redhat.com>

* add unit tests

and any changes necessary in relation to adding unit tests.

Signed-off-by: Bryce Palmer <bpalmer@redhat.com>

* verify schemas and fail if they don't match

Signed-off-by: Bryce Palmer <bpalmer@redhat.com>

* add better validations and decrease some code duplication

as well as add the necessary additional test cases

Signed-off-by: Bryce Palmer <bpalmer@redhat.com>

* remove some TODO comments and hardcode containerTool as 'docker'

as per review comments

Signed-off-by: Bryce Palmer <bpalmer@redhat.com>

* update the custom veneer builder

to work like the other veneer builders by reading a full FBC from STDOUT
and writing it to the output destination in the custom veneer config.
Also adds additional test cases to cover the new changes.

Signed-off-by: Bryce Palmer <bpalmer@redhat.com>

Signed-off-by: Bryce Palmer <bpalmer@redhat.com>