remove 'automatically generated' strings from repo
Maintainers use an import organizer called 'gci' that ignores files with
the string 'automatically generated' when called with its
--skip-generated flag.

This commit is the result of the following command:

```
rg ' automatically generated' -l | xargs -n1 sed -i 's/ automatically generated//g'
```

None of the changed files are _actually_ generated files. We do not want
gci to think they are.

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>